[GUI] Put error into separate module
