+++ owmods_gui/frontend/src/types.d.ts
@@ -121,9 +121,38 @@ export interface ModPrerelease {
     version: string;
 }
 
-/** Contains URL for a mod's thumbnail */
+/**
+ * Contains URL for a mod's thumbnail
+ *
+ * Note these paths are relative to the database website:
+ *
+ * `https://ow-mods.github.io/ow-mod-db/thumbails/`
+ *
+ * This should be prepended to the URL to get the full URL.
+ *
+ * Also note that open_graph is always `None` for mods with a static thumbnail,
+ * so to always get a static thumbnail use `main` and `open_graph` together:
+ *
+ * ```
+ * # use owmods_core::mods::remote::ModThumbnail;
+ *
+ * let thumb = ModThumbnail {
+ * main: Some("main.gif".to_string()),
+ * open_graph: Some("open_graph.webp".to_string())
+ * };
+ *
+ * let animated = thumb.main.unwrap();
+ * let static_thumb = thumb.open_graph.unwrap_or(animated.clone());
+ *
+ * assert_eq!(animated, "main.gif");
+ * assert_eq!(static_thumb, "open_graph.webp");
+ * ```
+ *
+ */
 export interface ModThumbnail {
+    /** Main thumbnail, this will be animated if the mod has an animated thumbnail */
     main?: string;
+    /** Open-graph image, this will always be a static image but will always be `None` on mods with a static thumbnail */
     openGraph?: string;
 }
